vue-template-compiler as a peer dependency
`vue-template-compiler` is now a peer dependency instead of a direct dependency.
This allows the user to pin `vue-template-compiler` to a specific version instead of relying on the implicit upgrades from a semver caret range.

Without this change, updating Vue.js to version 2.4.2 can result in
errors like this:

Vue packages version mismatch:

- vue@2.4.2
- vue-template-compiler@2.4.1

This may cause things to work incorrectly. Make sure to use the same version for both.
If you are using vue-loader@>=10.0, simply update vue-template-compiler.
If you are using vue-loader@<10.0 or vueify, re-installing vue-loader/vueify should bump vue-template-compiler to the latest.

This was first reported in
[vue-loader](https://github.com/vuejs/vue-loader/issues/470) and later
[fixed in version
10.0.0](https://gist.github.com/yyx990803/faebe22e8763f5b17572b35ed96f52fe#vue-loader1000).

There is also [an open issue on
vueify](https://github.com/vuejs/vueify/issues/159) but it's not
currently fixed.