Bug fix so rollup-plugin-vue works with PostCSS
Apologies if this isn't a good fix. It works for me, but it's entirely possible there is a better solution. I seem to notice in older versions of the plugin, the css files has .css extensions and now they do not. Perhaps this was the source of the issue, or I am completely wrong and the files never had .css. Hope this helps.

https://github.com/vuejs/rollup-plugin-vue/issues/241