Possible Bug Fix
Not entirely sure why this bug works. But this is triggered from Rollup generating two packages, my boilterplate is used for building and distributing reusable Vue components. The first package is a UMD build or iife, then the second is usually and es6 build. rollup-plugin-vue works great on the first pass, even if I change the output order. The second pass gives this error:

Github/vue-giveworks-form/src/main.js → dist/vue-giveworks-form.es.js...
51% (199): src/Plugins/GiveworksForm.js[!] (VuePlugin plugin) Error: TypeError: aStr.charAt is not a function
src/Components/GiveworksForm.vue
Error: TypeError: aStr.charAt is not a function
    at Object.<anonymous> (/Users/justinkimbrell/Github/vue-giveworks-form/node_modules/rollup-plugin-vue/dist/index.js:117:35)
    at Generator.next (<anonymous>)
    at fulfilled (/Users/justinkimbrell/Github/vue-giveworks-form/node_modules/rollup-plugin-vue/dist/index.js:4:58)

The style.map.mappings is a string on the first pass. It is an array on the second pass and it throws that error above. There is probably a better way to fix this, but it’s unclear as to what style.map.mapping is used for and what its a string or an array.